Seguimiento de alta de solicitud de permiso

+++ IICA/Models/Entidades/Result.cs
@@ -10,7 +10,7 @@ namespace IICA.Models.Entidades
         public bool status { get; set; }
         public string mensaje { get; set; }
         public string codigoHttp { get; set; }
-        public int id { get; set; }//utilizado para almacenar un id de algun registro q se  acabe de registrar o actualizar
+        public Int64 id { get; set; }//utilizado para almacenar un id de algun registro q se  acabe de registrar o actualizar
         public Object objeto { get; set; }
     }
 }
